Fix pre check
